Create WEEK 6 (Question 2)
Ques2:Given a graph, design an algorithm and implement it using a program to find if a graph is
bipartite or not. (Hint: use BFS)

Input Format: 
Input will be the graph in the form of adjacency matrix or adjacency list. 

Output Format: 
Output will be 'Yes Bipartite' if graph is bipartite, otherwise print 'Not Bipartite'