Create WEEK 8 (Question 3)
Ques 3:Assume that same road construction project is given to another person. The amount he will 
earn from this project is directly proportional to the budget of the project. This person is 
greedy, so he decided to maximize the budget by constructing those roads who have highest 
construction cost. Design an algorithm and implement it using a program to find the maximum 
budget required for the project.
Input Format: 
The first line of input takes number of vertices in the graph. 
Input will be the graph in the form of adjacency matrix or adjacency list. 
Output Format: 
Out will be maximum spanning weight