Create WEEK 10 (Question 1)
Ques1: Given a list of activities with their starting time and finishing time. Your goal is to 
select maximum number of activities that can be performed by a single person such that selected
activities must be non-conflicting. Any activity is said to be non-conflicting if starting time 
of an activity is greater than or equal to the finishing time of the other activity. Assume that 
a person can only work on a single activity at a time.

Input Format: 
First line of input will take number of activities N. 
Second line will take N space-separated values defining starting time for all the N activities. 
Third line of input will take N space-separated values defining finishing time for all the N
activities.

Output Format: 
Output will be the number of non-conflicting activities and the list of selected activities.