Create WEEK 11 (Question 1)
Ques1:Given a sequence of matrices, write an algorithm to find most efficient way to multiply 
these matrices together. To find the optimal solution, you need to find the order in which these
matrices should be multiplied. 
Input Format: 
First line of input will take number of matrices n that you need to multiply. 
For each line i in n, take two inputs which will represent dimensions aXb of matrix i.
Output Format: 
Output will be the minimum number of operations that are required to multiply the list of
matrices