Create WEEK 11 (Question 2)
Ques2: Given a set of available types of coins. Let suppose you have infinite supply of each 
type of coin.
For a given value N, you have to Design an algorithm and implement it using a program to find
number of ways in which these coins can be added to make sum value equals to N.
Input Format: 
First line of input will take number of coins that are available. 
Second line of input will take the value of each coin. 
Third line of input will take the value N for which you need to find sum.
Output Format: 
Output will be the number of ways.