Create WEEK 10 (Question 2)
Given a long list of tasks. Each task take specific time to accomplish it and each task has a deadline associated with it. You have to  design an algorithm and implement it using a program to find maximum number of tasks that can be completed without crossing their deadlines and also find list of selected tasks.